Migrate *UpdateCoordinator to its own file.
